fixed #23 main.js とsphero-controller.js を統合
