+++ main.js
@@ -2,30 +2,49 @@ var express = require("express");
 var app = express();
 var http = require("http").Server(app);
 var io = require("socket.io")(http);
-var controller = require("./sphero-controller");
 
-var angles = [];
 // 自分の Sphero の ID に置き換える
-var port = "COM7";
+var orb = require("sphero")("COM7");
+var angles = [];
 var currentAngleIndex = 0;
 var isRunning = false;
 
 // 接続された時に呼び出されます。
 function onConnect() {
-  controller.setColor("orange");
+  orb.color("orange");
 }
 
 // 衝突時に呼び出されます。
 function onCollide(count) {
   console.log(isRunning);
   if (isRunning) {
-    controller.move(100, angles[Math.min(currentAngleIndex++, angles.length - 1)]);
+    roll(100, angles[Math.min(currentAngleIndex++, angles.length - 1)]);
   }
 }
 
-controller.connect(port, onConnect);
-controller.addEventListener("collision", onCollide);
+orb.connect(onConnect);
+orb.detectCollisions();
+orb.on("collision", onCollide);
 
+function roll(speed, degree) {
+  clearTimeout(moveLoopId);
+  orb.roll(speed, degree);
+  moveLoopId = setTimeout(roll, 1000, speed, degree);
+}
+function setColor(color, time) {
+  orb.getColor(function (err, data) {
+    if (data) {
+      // なぜかdata.colorは、16進数だが文字列として帰ってくるので、parseInt。
+      var originalColor = parseInt(data.color);
+      orb.color(color);
+      if (typeof time !== "undefined") {
+        setTimeout(function () {
+          orb.color(originalColor);
+        }, time * 1000);
+      }
+    }
+  });
+}
 app.use(express.static("client"));
 http.listen(3000, function () { });
 
@@ -42,21 +61,21 @@ io.sockets.on("connection", function (socket) {
     if (!isError) {
       currentAngleIndex = 0;
       if (angles.length > 0) {
-        controller.move(100, angles[currentAngleIndex++]);
+        roll(100, angles[currentAngleIndex++]);
       }
       isRunning = true;
     }
   });
-  socket.on("stopSphero", function() {
-    controller.move(0, 0);
+  socket.on("stopSphero", function () {
+    roll(0, 0)
     isRunning = false;
     currentAngleIndex = 0;
   });
-  socket.on("startCalibration", function() {
-    controller.startCalibration();
+  socket.on("startCalibration", function () {
+    orb.startCalibration();
   });
-  socket.on("finishCalibration", function() {
-    controller.finishCalibration();
+  socket.on("finishCalibration", function () {
+    orb.finishCalibration();
   });
 });
